Fix ruff lint findings and apply ruff format
Combines (a) ruff --fix auto-fixes, (b) ruff format normalization,
and (c) manual fixes for rules ruff could not auto-correct.

Manual fixes:
- B904 (raise from) in __main__.py x2
- PERF401/403 (list/dict comprehensions) in checks.py, invoke_pi.py, publish.py
- PIE810 (combined startswith) in hunks.py
- RUF046 (redundant int(round(...))) in progress.py x2
- RUF034 (always-APPROVE branch) in preflight.py — suppressed with noqa+comment;
  both branches return APPROVE intentionally as a placeholder for future
  APPROVE_WITH_NITS semantics
- E501 line-too-long in publish.py (string split)
- SIM105 (contextlib.suppress) in publish.py x2
- SIM103 (return condition) in publish.py
- F841 (unused var nit_in_scale) removed in render.py
- PT018 (composite assert) split in test_docs_consistency.py
- PT019 (_scrubbed_env fixture) -> @pytest.mark.usefixtures in test_invoke_pi.py x15

Configuration:
- Add tools.ruff.extend-exclude=["tests/adversarial"]; adversarial tests are
  frozen fixtures and must not be auto-formatted.

+++ momus/progress.py
@@ -105,13 +105,13 @@ class ProgressTracker:
         return min(total, 1.0)
 
     def percent(self) -> int:
-        return int(round(self.fraction() * 100))
+        return round(self.fraction() * 100)
 
     def bar(self, width: int = 20) -> str:
         """Render a unicode progress bar of the given character width."""
         if width <= 0:
             return ""
-        filled = int(round(self.fraction() * width))
+        filled = round(self.fraction() * width)
         filled = max(0, min(filled, width))
         return "█" * filled + "░" * (width - filled)
